backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
channels==4.0.0
daphne==4.0.0
Pillow==10.1.0
django-filter==23.3
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
